refactor: audit and harden 58 files across SDK (R132-R231)
Systematic code quality audit covering unwrap elimination, clippy fixes,
type safety improvements, and dead code removal across all major modules.

Key changes:
- Eliminate unsafe .unwrap()/.expect() in production code paths
- Fix clone_on_copy: add Copy to CacheConfig, remove redundant .clone()
- Type safety: i32→i64 for neo_token votes/gas, u64 as i64→try_from
- Fix PartialEq bug in ProviderError (7 missing match arms)
- Fix unreachable match arms in invocation_result deserializer
- Replace expect_throw with recoverable error in wasm ws/backend
- Simplify invoke_function match block, remove needless return
- Use matches!() macro, .values() iterator, into_owned() idioms
- Remove unused imports in monitoring/metrics and monitoring/tracing
- Add Default impl for SecureStorage, fix non-existent SgxError variant
- Safer float-to-u64 ceiling, lossless u32→u64 via From trait

+++ src/neo_contract/traits/fungible_token.rs
@@ -14,16 +14,16 @@ pub trait FungibleTokenTrait<'a, P: JsonRpcProvider>: TokenTrait<'a, P> {
 	const BALANCE_OF: &'static str = "balanceOf";
 	const TRANSFER: &'static str = "transfer";
 
-	async fn get_balance_of(&self, script_hash: &ScriptHash) -> Result<i32, ContractError> {
+	async fn get_balance_of(&self, script_hash: &ScriptHash) -> Result<i64, ContractError> {
 		self.get_balance_of_hash160(script_hash).await
 	}
 
-	async fn get_balance_of_hash160(&self, script_hash: &H160) -> Result<i32, ContractError> {
+	async fn get_balance_of_hash160(&self, script_hash: &H160) -> Result<i64, ContractError> {
 		self.call_function_returning_int(Self::BALANCE_OF, vec![script_hash.into()])
 			.await
 	}
 
-	async fn get_total_balance(&self, wallet: &Wallet) -> Result<i32, ContractError> {
+	async fn get_total_balance(&self, wallet: &Wallet) -> Result<i64, ContractError> {
 		let mut sum = 0;
 		for account in wallet.accounts.values() {
 			sum += self.get_balance_of(&account.address_or_scripthash().script_hash()).await?;
@@ -35,7 +35,7 @@ pub trait FungibleTokenTrait<'a, P: JsonRpcProvider>: TokenTrait<'a, P> {
 		&self,
 		from: &Account,
 		to: &ScriptHash,
-		amount: i32,
+		amount: i64,
 		data: Option<ContractParameter>,
 	) -> Result<TransactionBuilder<P>, ContractError> {
 		let mut builder = self
@@ -54,7 +54,7 @@ pub trait FungibleTokenTrait<'a, P: JsonRpcProvider>: TokenTrait<'a, P> {
 		&self,
 		from: &ScriptHash,
 		to: &ScriptHash,
-		amount: i32,
+		amount: i64,
 		data: Option<ContractParameter>,
 	) -> Result<TransactionBuilder<P>, ContractError> {
 		if amount < 0 {
@@ -73,7 +73,7 @@ pub trait FungibleTokenTrait<'a, P: JsonRpcProvider>: TokenTrait<'a, P> {
 		&self,
 		from: &ScriptHash,
 		to: &ScriptHash,
-		amount: i32,
+		amount: i64,
 		data: Option<ContractParameter>,
 	) -> Result<Bytes, ContractError> {
 		self.build_invoke_function_script(
@@ -94,7 +94,7 @@ pub trait FungibleTokenTrait<'a, P: JsonRpcProvider>: TokenTrait<'a, P> {
 		&self,
 		from: &Account,
 		to: &NNSName,
-		amount: i32,
+		amount: i64,
 		data: Option<ContractParameter>,
 	) -> Result<TransactionBuilder<P>, ContractError> {
 		let mut builder = self
@@ -113,7 +113,7 @@ pub trait FungibleTokenTrait<'a, P: JsonRpcProvider>: TokenTrait<'a, P> {
 		&self,
 		from: &ScriptHash,
 		to: &NNSName,
-		amount: i32,
+		amount: i64,
 		data: Option<ContractParameter>,
 	) -> Result<TransactionBuilder<P>, ContractError> {
 		let script_hash = self.resolve_nns_text_record(to).await?;
